requirements.txt
easydict==1.10
h5py==2.10.0
jsonlines==2.0.0
lmdb==1.4.1
more_itertools==10.1.0
msgpack_numpy==0.4.8
msgpack_python==0.5.6
networkx==2.5.1
numpy==1.22.3
opencv_python==4.7.0.72
Pillow==10.1.0
progressbar33==2.4
psutil==5.9.4
PyYAML==6.0.1
ray==2.8.0
requests==2.25.1
shapely==2.0.1
timm==0.9.2
torch==1.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.64.1
transformers==4.28.0
